// src/components/ui/sidebar/types.ts
export enum SidebarState {
  Expanded = 'EXPANDED',
  Collapsed = 'COLLAPSED',
}

export enum SidebarSide {
  Left = 'LEFT',
  Right = 'RIGHT',
}

export enum SidebarVariant {
  Sidebar = 'SIDEBAR',
  Floating = 'FLOATING',
  Inset = 'INSET',
}

export enum SidebarCollapsible {
  Offcanvas = 'OFFCANVAS',
  Icon = 'ICON',
  None = 'NONE',
}